requirements.txt
playwright
pytest
allure-pytest
requests

